deprecated ShareLinkBuilder, moved it out of Branch class and renamed as BranchShareSheetBuilder
